move check_treestatus into util
